Inline store conformance tests
