+++ packages/test-utils/src/index.ts
@@ -1,1 +1,2 @@
 export { createTestSdkFactory } from "./workflow-runner";
+export { testSchema } from "./schema";
