+++ packages/test-utils/src/schema.ts
@@ -1,0 +1,13 @@
+import type { StandardSchemaV1 } from "@yieldstar/core";
+
+export function testSchema<T>(): StandardSchemaV1<unknown, T> {
+  return {
+    "~standard": {
+      version: 1,
+      vendor: "yieldstar-test",
+      validate(value) {
+        return { value: value as T };
+      },
+    },
+  };
+}
